@@ -12,43 +12,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.4",
-    "@angular/common": "^4.4.4",
-    "@angular/compiler": "^4.4.4",
-    "@angular/core": "^4.4.4",
-    "@angular/forms": "^4.4.4",
-    "@angular/http": "^4.4.4",
-    "@angular/platform-browser": "^4.4.4",
-    "@angular/platform-browser-dynamic": "^4.4.4",
-    "@angular/router": "^4.4.4",
+    "@angular/animations": "^5.0.2",
+    "@angular/common": "^5.0.2",
+    "@angular/compiler": "^5.0.2",
+    "@angular/core": "^5.0.2",
+    "@angular/forms": "^5.0.2",
+    "@angular/http": "^5.0.2",
+    "@angular/platform-browser": "^5.0.2",
+    "@angular/platform-browser-dynamic": "^5.0.2",
+    "@angular/router": "^5.0.2",
     "angular2-flash-messages": "^1.0.8",
     "angular2-jwt": "^0.2.3",
-    "bulma-checkradio": "0.0.9",
-    "core-js": "^2.4.1",
+    "bulma-checkradio": "^0.1.0",
+    "core-js": "^2.5.1",
     "jwt-decode": "^2.2.0",
-    "rxjs": "^5.4.2",
+    "rxjs": "^5.5.2",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "^1.4.7",
-    "@angular/compiler-cli": "^4.4.5",
-    "@angular/language-service": "^4.4.5",
-    "@types/jasmine": "~2.6.0",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^8.0.46",
-    "codelyzer": "~3.2.1",
+    "@angular/cli": "^1.5.2",
+    "@angular/compiler-cli": "^5.0.2",
+    "@angular/language-service": "^5.0.2",
+    "@types/jasmine": "~2.8.2",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^8.0.53",
+    "codelyzer": "~4.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^1.7.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.2.0",
     "ts-node": "~3.3.0",
     "tslint": "~5.8.0",
-    "typescript": "~2.5.3",
-    "webpack": "^3.7.1"
+    "typescript": "~2.6.1",
+    "webpack": "^3.8.1"
   }
 }
