@@ -2,7 +2,6 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
 import { HttpClientModule } from '@angular/common/http';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { JwtHelper } from 'angular2-jwt';
@@ -17,9 +16,13 @@ import { CharactersListComponent } from './characters/components/characters-list
 import { NewCharacterComponent } from './characters/components/new-character.component';
 import { CharacterComponent } from './characters/components/character.component';
 
-import { CharactersSidebarComponent } from './lobbies/components/characters-sidebar.component';
-import { ResearchListComponent } from './lobbies/components/research-list.component';
+import { CharacterSidebarComponent } from './lobbies/components/character-sidebar.component';
+import { ListLobbyComponent } from './lobbies/components/list-lobby.component';
+import { NewLobbyComponent } from './lobbies/components/new-lobby.component';
 import { LobbyComponent } from './lobbies/components/lobby.component';
+import { DisplayComponentService } from './lobbies/display-component.service';
+
+import { GroupComponent } from './groups/components/group.component';
 
 import { NavbarComponent } from './shared/components/navbar/navbar.component';
 import { FooterComponent } from './shared/components/footer/footer.component';
@@ -28,6 +31,7 @@ import { AuthService } from './services/auth.service';
 import { CharacterService } from './services/character.service';
 import { RegisterService } from './services/register.service';
 import { AuthGuardService as AuthGuard } from './services/auth-guard.service';
+import { GroupGuardService as GroupGuard } from './services/group-guard.service';
 import { LobbyService } from './services/lobby.service';
 
 import { TokenInterceptor } from './services/interceptors/token.interceptor';
@@ -36,7 +40,8 @@ const appRoutes: Routes = [
   {path: '', component:LoginComponent},
   {path: 'register', component:RegisterComponent},
   {path: 'players/:id/characters', component:CharacterComponent, canActivate: [AuthGuard]},
-  {path: 'players/:id/search', component:LobbyComponent, canActivate: [AuthGuard]}
+  {path: 'players/:id/search', component:LobbyComponent, canActivate: [AuthGuard, GroupGuard]},
+  {path: 'groups/:id', component:GroupComponent, canActivate: [AuthGuard]}
 ]
 
 @NgModule({
@@ -46,15 +51,17 @@ const appRoutes: Routes = [
     LoginComponent,
     RegisterComponent,
     CharacterComponent, CharactersListComponent, NewCharacterComponent,
-    LobbyComponent, CharactersSidebarComponent, ResearchListComponent
+    LobbyComponent, CharacterSidebarComponent, ListLobbyComponent, NewLobbyComponent,
+    GroupComponent
   ],
   imports: [
     BrowserModule, FormsModule, ReactiveFormsModule,
-    HttpModule, HttpClientModule, RouterModule.forRoot(appRoutes)
+    HttpClientModule, RouterModule.forRoot(appRoutes)
   ],
   providers: [
     AuthService, RegisterService, CharacterService, LobbyService,
-    AuthGuard,
+    DisplayComponentService,
+    AuthGuard, GroupGuard,
     {provide: HTTP_INTERCEPTORS, useClass: TokenInterceptor, multi: true},
     JwtHelper
   ],
